+++ examples/state_with_data.rs
@@ -31,5 +31,5 @@ fn main() {
     let mut sm = StateMachine::new(Context);
     let result = sm.process_event(Events::Event1);
 
-    assert!(result == Ok(&States::State2(MyStateData(42))));
+    assert!(matches!(result, Ok(&States::State2(MyStateData(42)))));
 }