+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Starting state can now contain data ([issue-34](https://github.com/korken89/smlang-rs/issues/34))
 
+### Changed
+- Custom guard error types are now specified as a type of the `StateMachineContext` to allow for
+  more versatile types.
+
 ## [v0.5.1]
 
 ### Fixed